+++ spec/rubocop/cop/style/def_without_parentheses_spec.rb
@@ -6,27 +6,32 @@ module Rubocop
   module Cop
     module Style
       describe DefWithoutParentheses do
-        let(:def_par) { DefWithoutParentheses.new }
+        let(:cop) { described_class.new }
 
         it 'reports an offence for def with parameters but no parens' do
           src = ['def func a, b',
                  'end']
-          inspect_source(def_par, src)
-          expect(def_par.offences.size).to eq(1)
+          inspect_source(cop, src)
+          expect(cop.offences.size).to eq(1)
         end
 
         it 'reports an offence for class def with parameters but no parens' do
           src = ['def Test.func a, b',
                  'end']
-          inspect_source(def_par, src)
-          expect(def_par.offences.size).to eq(1)
+          inspect_source(cop, src)
+          expect(cop.offences.size).to eq(1)
         end
 
         it 'accepts def with no args and no parens' do
           src = ['def func',
                  'end']
-          inspect_source(def_par, src)
-          expect(def_par.offences).to be_empty
+          inspect_source(cop, src)
+          expect(cop.offences).to be_empty
+        end
+
+        it 'auto-adds required parens' do
+          new_source = autocorrect_source(cop, 'def test param; end')
+          expect(new_source).to eq('def test (param); end')
         end
       end
     end